requirements.txt
Django==1.9
line-bot-sdk==1.5.0
gunicorn==19.3.0
pandas==0.25.0
bs4==0.0.1
lxml==4.3.4
beautifulsoup4==4.8.0
numpy==1.16.4
